// src/lib/api/generated/models/getSellerReviewsSortBy.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Marketplace API
 * API for Village Website & Marketplace
 * OpenAPI spec version: 1.0.0
 */

/**
 *
 */
export type GetSellerReviewsSortBy = typeof GetSellerReviewsSortBy[keyof typeof GetSellerReviewsSortBy];


export const GetSellerReviewsSortBy = {
  createdAt: 'createdAt',
  rating: 'rating',
} as const;