// src/lib/api/generated/models/getSellerReviewsSortOrder.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Marketplace API
 * API for Village Website & Marketplace
 * OpenAPI spec version: 1.0.0
 */

export type GetSellerReviewsSortOrder = typeof GetSellerReviewsSortOrder[keyof typeof GetSellerReviewsSortOrder];


export const GetSellerReviewsSortOrder = {
  asc: 'asc',
  desc: 'desc',
} as const;